backend/ecommerce_backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cloudinary==1.37.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
kafka-python==2.0.2
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
djangorestframework-simplejwt==5.3.1